Some refactoring and stage caching tested
